requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
discord.py==2.3.2
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
tabulate==0.9.0
tqdm==4.66.1
tzdata==2023.3
yarl==1.9.2
pymongo==4.5.0
pytest==7.4.2
pytest-asyncio==0.21.1
